DOC: deduplicate stats.power.normal_power docs
Prior to this commit, the extended summaries of the
`stats.power.normal_power` and `stats.power_normal_power_het`
functions are the same. These methods use different assumptions, so
the extended summary of `stats.power.normal_power` is changed to
indicate that the population variance of the null and alternative
hypotheses are the same.

The short summary of both of the aforementioned methods is slightly
altered for grammatical purposes.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>